map items and rhs filter
added map items where an item which has a string inside can be mapped to that string (e.g. item23_232423498 => item32) which is useful to group by items from the same category.

also added rhs filter to only include rules that have a specific string in their rhs (e.g. item23) which is useful when we are targeting a particular group of events in the rhs.